CRM457-947: System specs (#254)
## Description of change

- Added ~15 system specs so that the total line coverage was over 80%,
and for most controllers line coverage was over 90% and all major views
exercised
- Fixed some bugs that came up in the process
- Refactored how downloads work for improved developer ergonomics and a
minor security improvement

## Link to relevant ticket

[JIRA](https://dsdmoj.atlassian.net/browse/CRM457-947)

## Notes for reviewer
We were generating temporary download links any time we rendered the
supporting evidence screen, instead of waiting until the user clicked on
a link to a file. This meant we had to leave the links valid for 15
minutes, and in a dev environment that whole screen was inaccessible if
you didn't have S3 set up. So I did a quick shift so that temporary
links are now only generated a the point of click.

---------

Co-authored-by: Ivan Yohuno <ivan.yohuno@digital.justice.gov.uk>