CRM457: Add javascript for service and travel costs
Provide button that, when clicked, recalculate the value
as the adjusted cost that will be save unless they cancel.

This only loads the JS if the applicable container exists on
the page, to avoid unnecessary loading and possible clashes.

This also disable turbo frames in the button link to the forms
as this was preventing the form page  from loading the JS until
it was hard refreshed. :(