CRM457-1790: Discard  all adjustments prior to sending back (#646)
## Description of change
If I make adjustments then send an application back, those adjustments
are effectively ignored, in that the provider app does not pick them up
and when the provider app pushes a new version to the app store, those
adjustments are overwritten anyway. BUT there's an unintended side
effect, which is that when trying to work out which fields the provider
has changed, the provider app diffs its database with the latest app
store version. It picks up on the adjustments in the diff, with the end
result that in the caseworker UI the fields the caseworker previously
adjusted end up being highlighted as ones the _provider_ has adjusted in
the RFI loop.

This PR undoes caseworker adjustments when an application is sent back,
so that the version available in the app store has no adjustments, so
the "what has the provider changed?" diff no longer gets messed with.

[Link to relevant
ticket](https://dsdmoj.atlassian.net/browse/CRM457-1790)