CRM457-2067: Sync risk changes back to app store (#782)
## Description of change
Let the app store know immediately if the risk has changed.

[Link to relevant
ticket](https://dsdmoj.atlassian.net/browse/CRM457-2067)

## Notes for reviewer
I opted to make the app store call synchronous. As we move towards the
app store as our single source of truth we need to get comfortable with
the idea of it being a direct dependency that we interact with as part
of our request/response interaction with end users. A synchronous
request means we guarantee things stay in sync, get immediate feedback
if there is a problem, and start to validate that our architecture can
handle using the app store as the single source of truth.