CRM457-2049: Fix RFI/'delete all adjustments' edge case (#797)
This reverts commit 892d599b033366fe755961c33916c9c4b211838d.

Now when we click 'delete all adjustments' we manually remove every
adjustment from the payload instead of pulling a 'fresh copy' from the
app store. This is because the app store version may contain adjustments
if we have completed an RFI loop and adjustments were made before the
claim was sent back.

Robin had written code to do it this way initially, and I (wrongly!)
said it was much simpler to take the 'fresh copy' approach.