CRM457-2155: Exclude high value claims from auto-assignment (#824)
## Description of change
Note that there are still 3 open claims that don't have a cost summary
in the blob as they precede that. They are all high risk claims that
have always in the past been allocated manually, so they should continue
to be done that way. I've added in some explicit fallback logic to
describe that behaviour.

[Link to relevant
ticket](https://dsdmoj.atlassian.net/browse/CRM457-2155)