Fix disbursement change edge case (#852)
## Description of change
If a mileage-based disbursement has had the number of miles changed,
then `total_cost_without_vat` will not change, as now that is only used
to represent lump-sum disbursements, and we don't put a calculated total
for miles * rate in there any more. So the only way to be confident that
there has been a chance is to compare the total, vat-inclusive
calculated result.