Memoize data_for_calculation (#876)
CoreCostSummary.show_allowed? calls `changed?` on a view model for every
single work item and disbursement (can be hundreds), and each of those
calls a calculator method with `submission.data_for_calculation` as an
argument. This is a slow method when there are hundreds of work items
and disbursements, so we can get a quick win by memoizing it so it only
runs once per page load.